tests: extension registration unit testing port
The following ports over the extension registration test from the legacy
container into the unit testing root.
